Address repeated page review comments
