+++ modules/parquet/src/parquetjs/parser/parquet-reader.ts
@@ -371,7 +371,7 @@ export class ParquetReader {
     return await decodeDataPages(pagesBuf, {...context, dictionary});
   }
 
-  /** Reads and decodes contiguous data pages that begin and end on complete row boundaries. */
+  /** Reads and decodes only the contiguous data pages overlapping one non-repeated row range. */
   async readColumnChunkRange(
     schema: ParquetSchema,
     columnChunk: ColumnChunk,
@@ -384,6 +384,9 @@ export class ParquetReader {
     }
     const columnMetadata = columnChunk.meta_data!;
     const field = schema.findField(columnMetadata.path_in_schema);
+    if (field.repetitionType === 'REPEATED' || field.rLevelMax !== 0) {
+      throw new Error('Selective Parquet page reads currently require non-repeated columns');
+    }
     const type: PrimitiveType = getThriftEnum(Type, columnMetadata.type) as any;
     if (type !== field.primitiveType) {
       throw new Error(`chunk type not matching schema: ${type}`);
@@ -406,12 +409,7 @@ export class ParquetReader {
       dLevelMax: field.dLevelMax,
       compression,
       column: field,
-      // Repeated leaves have one level entry per physical value, not one per logical row. Let the
-      // page decoder consume the selected page range and preserve all repetition levels.
-      numValues:
-        field.rLevelMax === 0
-          ? new CompactInt64(lastPage.endRowIndex - firstPage.firstRowIndex)
-          : undefined,
+      numValues: new CompactInt64(lastPage.endRowIndex - firstPage.firstRowIndex),
       dictionary: [],
       preserveBinary: this.props.preserveBinary,
       retainByteArrayViews: this.props.retainByteArrayViews,
@@ -433,9 +431,6 @@ export class ParquetReader {
       await this.file.read(firstPage.offset, dataLength, signal ?? this.props.signal)
     );
     const decoded = await decodeDataPages(dataBuffer, {...context, dictionary});
-    if (field.rLevelMax !== 0) {
-      return decoded;
-    }
     const relativeStart = rowRange.start - firstPage.firstRowIndex;
     const relativeEnd = relativeStart + rowRange.end - rowRange.start;
     return sliceNonRepeatedColumnChunk(decoded, field.dLevelMax, relativeStart, relativeEnd);
